indexing/requirements.txt
chromadb >=0.4.0,<0.5.0
transformers >=4.0.0,<5.0.0
gitpython >=3.0.0,<4.0.0
python-dotenv
numpy<2
sentencepiece 
joblib 
requests 